requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
email-validator>=2.2.0
exceptiongroup==1.2.2
Flask>=3.1.0
Flask-Login>=0.6.3
Flask-SQLAlchemy==3.0.5
Flask-WTF>=1.2.2
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-speech==2.32.0
google-cloud-texttospeech==2.26.0
googleapis-common-protos==1.70.0
greenlet==3.2.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn>=23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe>=3.0.2
openai>=1.75.0
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary>=2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.23
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uv==0.6.14
Werkzeug>=3.1.3
WTForms>=3.2.1
